requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
pytest==7.0.0
selenium==3.141.0